[symbolizer] Add llvm-symbolizer-restart-interval flag
In a previous change
(https://fuchsia-review.googlesource.com/c/fuchsia/+/342170), we
added code to restart the llvm-symbolizer on every query. This succeeded
in reducing memory consumption, but increased running time greatly for
some use cases. This flag will let us tune the behavior to balance
memory usage vs running time for different use cases.

In this change the flag defaults to 1 to prserve the existing behavior.
Once this is merged and rolled, we can add code to the infrastructure
recipes to set this flag when invoking symbolizer, then we can change
the default to something more appropriate for non-infrastructure
use cases.

Bug: 42593
Change-Id: I28f820b72cc259c616b2b5473023f002710373a9