[inspect_deprecated] Fix crash caused by incorrect validation.
A bug caused us to not validate the result of listing children in
inspect FIDL reading. This caused the result in question to be unwrapped
even though it was empty and crash.

This change fixes the problem and provides a minor refactor of a couple
methods in the deprecated API to simplify testing this case.

Fixed: 42199
Fixed: 42439
Fixed: 42023
Change-Id: I0538ab6211c57518ae5f415b46292a85dd9efb5b