[pkg-resolver] Add repo last used time property
This change adds a new inspect property to TUF repositories to track the
last time a repository was used, and modifies the semantics of num
packages fetched to only count targets with valid metadata.

As tuf repositories are always updated before use, the last_used_time
will currently always be shortly after the last_updated_time, but that
will soon change.

Bug: 22802
Change-Id: I6d14e96e9063927b99dc420b76fd460db8548ee1