[kernel][x86] Spectre V2: Add function to overwrite RAS/RSB
Spectre V2 is a family of speculative execution information leak
attacks that rely on hostile code polluting / poisoning indirect
branch predictor structures and then invoking 'victim' code and
causing it to consume poisoned indirect branch predictor state.

Return instructions on most x86 processors use a structure
known as a 'return address stack' to speculatively provide a
target address in advance of the actual computed target address
being available, as a restricted form of branch target predictor.

For the RAS/RSB to be used safely by 'victim', code, it must not
underflow and it must be filled with known-safe addresses.
Execute a sequence of 32 CALLs without RETs to fill up the RAS,
pointing each return at a 'speculation trap', an infinite loop
with a dispatch serializing instruction.

A follow-on CL will add this sequence to:
* User->kernel entries
* On #VMEXIT

See:
* AMD Software Techniques for Managing Speculation on AMD Processors
  Mitigation V2-3 suggests doing this
* Intel Deep Dive: Indirect Branch Restricted Speculation
  'RSB Overwrite sequence'
* Intel: Speculative Execution Side Channel Mitigations Revision 1.0
  "An RSB overwrite sequence is a sequence of instructions that
  includes 32 more near CALL instructions with non-zero displacements
  than it has near RETs."
* Spectre Returns! Speculation Attacks using the Return Stack Buffer
  (Koruyeh et al.); this paper suggests we need to unconditionally
  execute the RSB overwrite sequence, not just on Skylake+ CPUs

Bug: 12540 Speculative Execution Mitigations.
Bug: 33667 Spectre mitigations?

Change-Id: Id15bcf1f2ed81124121ec9453efe3ca72a916a22