[netstack][bridge] respect WritePacket assumption
This adds an IPv6 test to the existing bridge test, which revealed that
`(stack.NetworkDispatcher).DeliverNetworkPacket` sometimes mutates its
arguments (as in the case of IPv6). This means payloads must be copied
before being sent to each of a bridge's constituent endpoints, which is
also done in this change.

The latter change fixes issue 41521.

This partially reverts commit 94be92d34a5c6e586f37f457dc13728161bca0c9.

Fixed: 41521
Change-Id: Ia9c59e2760822deb629691422f6ac8926a068942