[process_builder] Improve handling of reservation VMAR
Improves how the reservation VMAR is destroyed, depending on whether
ProcessBuilder.build() otherwise succeeds or fails.

Also adds a couple new process_builder tests to validate that the VMO
passed to ProcessBuilder.set_vdso_vmo() is actually being used. This
indirectly tests that the ReservationVmar can be torn down as expected
if an error occurs during process creation, which eliminated a potential
theory for bug 42437.

Fixed: 42437
Change-Id: Idbe913aa4bf796123b75650946df4776c147595a