[pkgfs] Fix blobfs.HasBlob edge case
When importing a package meta far, pkgfs needs to determine which
package content blobs are already present in blobfs and which blobs need
to be fetched. The previous implementation assumed that if a blob could
not be created for write with an access denied error, the blob exists,
but that error case also includes blobs that are in the process of being
written.

This change modifies the method to consider a blob as being present iff
it can be opened for read and is readable.

Bug: 42189
Change-Id: I9f63372cc11121da0e97b34050920b05ad85ef86