[aml-uart] Fix race in HandleRX/TX()
It was previously possible for both ReadStateAndNotify() and
SerialImplAsyncReadAsync() to call HandleTX().  After one finished, the
other would write nothing and call the callback again, which wasn't
cleared.  This was theoretically possible with HandleRX() as well.

Move the check whether a read/write is pending into the HandleRX/TX()
functions, making them do nothing if there is no write pending.

As a safety measure, null the callbacks and cookies for the read and
write callbacks when they are to be called, so they will never be called
twice.

Fixed: 42413
Test: runtests -t aml-uart-test
runtests -t serial-async-unittest-test

Change-Id: Id05b2da716b49b5c92a41ed646f3d6590999d162