[pkgfs] Fix blob install edge case
When opening a /pkgfs/install/pkg/{blob} or /pkgfs/install/blob/{blob}
path, pkgfs was interpreting an access denied error from blobfs as a
signal that the blob already exists and is readable, but that error
could also mean that the blob is in the process of being written. As a
result, pkgfs could fulfill a need before that blob was present if
multiple clients concurrently write to the same blob.

This change adds an extra check to this path to differentiate between
blobs that are unwritable/readable and blobs that are
unwritable/unreadable.  In the latter case, pkgfs now bubbles up the
error to the caller without fulfilling the need.

Bug: 42189
Fixed: 39488
Change-Id: I2bc0a5ad71a4c5fc956f96c0cc6e81ba07f3799a