[soundplayer] call AudioRenderer.SetUsage once per renderer
AudioRenderer's behavior has changed such that its SetUsage method
may only be called prior to calling SetPcmStreamType. soundplayer
was calling that method every time it played a sound, possibly on a
renderer that was already used.

This CL makes usage a property of the renderer, sets usage when
creating a renderer, and discards the cached renderer if it has the
wrong usage.

TEST: The fake AudioRenderer in the integration test is modified to
      prohibit SetUsage after SetPcmStreamType. This makes the test
      fail without the other changes in this CL.
Change-Id: I9432fe15b169966e07972fc39698a6a6b881c28d