[camera] Require all sources to be enumerable in stream provider test
This change removes the allowance for stream providers to return null in
the stream_provider_test, in which case the test would be skipped. Now,
the test will fail if the provider cannot be created. Because the test
uses an injected camera manager instance that is torn down between runs,
it will not hold references to the camera controller upon exiting the
test.

The test will still fail if the global environment camera manager has
been started (e.g. by running the non-hermetic camera_demo) or by using
gtest_repeat, which does not tear down the test process between runs.
The latter is checked using a global counter, and fails with a
descriptive error message. The former is not detectible, and will
manifest as a failure to connect to the manager stream provider, as
the controller only supports a single client at a time.

Test: ran stream_provider_test with gtest_repeat=1, test passes
      ran stream_provider_test with gtest_repeat=10, test fails with
        expected error message
      ran camera_tests
Change-Id: If3b915c1751173257cd2492eb7207e4536183ab1