[SetUI][Intl] Scope ICU Data Loader to loading.
The ICU Data Loader must be created as it initializes the ICU data as
a side effect while it is in scope. However, the loader itself leads to over
10MB being allocated. We should only keep it in scope when necessary
(loading data).

This changelist reduces the scope of the loader.

Bug:  42254
Test: fx run-test setui_service_tests -- --test
Existing tests ensure the loader is properly initialized in the e2e flow.
If the loader was not present, the initialization will crash.

Change-Id: Ic24c5ceb6ae81acf15fe47236486ad7396c4ee1c