[description] Quote setting strings as necessary.
Setting values can have spaces in them which is handled by quoting in
the input handler. But when values were output there was no quoting or
escaping applied which meant that the user can't copy-and-paste the
strings and the help text was technically incorrect.

Adds a utility function to quote and escape strings for console
consumption, and use this in the settings formatter.

Change-Id: I3de4833248035add26eed53eba2a7cfe746de818