[sysmem] Permit reads of 0 from secure memory
For now at least, permit reads of 0 from secure mem, without any thread
death.  We previously attempted to write 0xF0 and flush cache, so a read
of 0 at least means the write, flush, read didn't work (good).

This may not be the ideal way to test this, but for now it avoids
failing the sysmem secure mem tests (observed on sherlock).

BUG: 42276
Change-Id: Ia57ab8be81d1a56df538d866fc4458d0cdc71885