[debugger][line input] Handle Control-C in options.
Adds the ability to set a Control-C handler for the line input class.

Use this handler for a modal option prompt. Callers can how specify
which option they would like to correspond to Control-C. By default it
has the previous behavior (clears line, asks again).

Hooks this up to the existing zxdb prompts.

Changes the zxdb quit prompt from "[y/n]" to "y/n >". This is because
the square-bracketed version looks too similar to the default "[zxdb]"
prompt rather than like a special modal question. I'm thinking we'll
standardize on ">" for modal prompts and "[]" for the regular one.

Change-Id: I4526843759ef947648acfb680c8af2fc37129661