[futex][ownership] Do not allow unstarted thread to own futexes.
Threads which have not started yet may not be the owner of a futex.
Enforce this at the futex context level of the kernel so that we do
not hit an ASSERT in the OwnedWaitQueue code which makes absolutely
sure that this cannot happen.

In addition, clarify the documentation around what happens if
++ A user attempts to assign ownership of a futex to a thread which
   has not started.
++ A user attempts to assign ownership of a futex to a thread which
   has already exited.
++ A thread exits while it owns a futex.

Finally, add formal tests for the first two clarifications on this
list.  The third (thread exits while it owns a futex) already had a
test.

Fixed: 41773
Change-Id: Idde77699810f514cd354d9ace47f9e2dc7f43093