[debugger] Step subexpression.
Prompts to step into a specific subexpression.

The command is added in its own file. This is a test of a new scheme.
The other commands will be converted to their own files in a separate
directory in a followup.

Add a new test harness that sets up more state for a typical console
test.

Change-Id: I36e781cdbe9b35e53fe9aaa680dcc149d2cf7b89