[msd-arm-mali] Hold lock around shutting down device in test
It's possible one thread's execution is delayed for a very long time and
it doesn't create its device before the device is starting to be
unbound. The client could handle that gracefully, but that would
increase the difficulty of detecting if the device failed to be created
for that reason or because of a bug. By putting the lock around the
entire teardown sequence we can ensure connections are created only when
that should be guaranteed to succeed.

Change-Id: I0aa37674d019b5bf4577d47944a7d46133476c11