[pkg] Add optional root_version to RepositoryConfig
The root keys in the RepositoryConfig need to correspond to the public keys
that signed the root metadata, and since the keys can be rotated, the only
metadata we can use to initialize trust before this patch is version 1.
This however carries the risk that if the root keys are ever compromised,
devices would still use those keys, which might lead an attacker to taking
over a system.

Instead, this patch lets us optionally specify the initial root metadata
version (with a default version of 1) we trust so that we can rotate keys
without worrying about breaking older clients.

Fixed: 41093
Change-Id: Ib6e875d087e9e0d848ae79361c971107cd667bfc