[cm] Split OutDir API of creating echo service / static files.
Previously, the call `add_echo_service` provided a set of default files
in the OutDir which are assumed to exist by several tests.

We modify the OutDir API to be more precise about what is being created,
moving the "create a default set of files" logic into the `RoutingTest`
object instead.

Change-Id: I6c358fc813cdd5e2ef0334676c702d4ec29776fa