[fidl][zircon] Fix early use of fidl coding tables
Previously, the FIDL coding tables were relying on being able to run
functions at global-constructor time in order to build themselves.

Because the tables are used by the C runtime library and code that runs
before main(), this fails if the compiler doesn't happen to do quite
extensive optimization and manage to get the whole object representation
of the tables into .rodata.

To fix this, the approach taken is to change lib/fidl/internal.h to
plain C structs, and use desginated initialization syntax to set up the
tables. This guarantees that they'll be in .data.rel.ro and will not
need any compiler-generated code to run before being usable.

One complexity is that the circularity of alt_type references
(previously resolved by using a function call) requires forward
declarations of static structures in the tables file. C allows this,
however C++ does not. This means that the tables file that's now using C
syntax must actually be compiled using a C compiler. This has some knock
on effects, e.g. in transformer_tests.cc, where the tables file was
previously being directly included into a .cc file, which is now not
possible.

This fixes a startup crash that was manifesting in the GCC build. But
it's UB on all compilers, so it was only a matter of time another change
caused a failure with another compiler or in another location.

Bug: 39978
Change-Id: I77fe36c4444be610ffee710889e1a717aeb73ca4