[fidlc] v1 coding tables are always xunions
fidlc would emit Unions and UnionPointers for both the old & v1 coding
tables. This was problematic for code that relied on the coding tables
for the v1 wire format, e.g. the FIDL validator, encoder & decoder
wouldn't work properly, since the coding tables would indicate that an
object is a union, instead of a xunion.

This CL modifies the coding tables so that:

1. An alt_type pointer is added to FidlCodedXUnion.

2. The alt_type pointer in a FidlCoded(Struct|Union|Array|Vector|XUnion)
   is now a pointer to a generic fidl_type_t, rather than being
   a pointer to a FidlCodedX. This is needed since a FidlCodedUnion can
   point to a FidlCodedXUnion (and vice versa).

Test: fx run-host-tests fidl_cpp_host_unittests
Test: fx run-host-tests fidl-compiler-test
Test: fx run-host-tests fidl-test
Test: fx build && fx qemu -kN && fx run-test fidl_tests

Change-Id: I8283b7f052b696a9acd52a34e923dbacd872b664