[ledger] fork lib/backoff into src/ledger
The exponential backoff util at lib/backoff is currently shared between
ledger, cobalt and appmgr. In preparation of Ledger move out of the
Fuchsia tree, this patch forks a copy of lib/backoff into src/ledger and
switches Ledger to use the private copy.

We also make the following changes to the backoff library for Ledger:

 - migrate to Ledger logging
 - remove FXL macros
 - change the namespace to ledger::

Change-Id: I0015ec1806e3472373839d66d9afdde8e60650ff