[brmcfmac] Move from watchdog polling to using interrupts
The brcmfmac driver used the watchdog to periodically poll for events.
This was because interrupts did not work properly in the past. As an
experiment I tried to enable them and they now seem to work properly.

This change removes the watchdog polling and enables the interrupt
thread that was previously disabled. The event handler used to have a
mutex to prevent the workqueue and watchdog from running the event
handler at the same time. This mutex did not work with the interrupt
thread however because it locks the func1 mutex and then waits for the
event handler mutex which might already be locked in the workqueue and
then the workqueue got stuck waiting for func1. This caused a deadlock
so this change removes the event handler mutex and replaces it with a
func1 mutex lock before entering the event handler. This works reliably
but may be a little heavy-handed, there might be room for improvement
here.

This affects every part of the drivers communication with firmware so
any major failures should be immediately obvious. I have tested it
extensively and it seems fine but thorough testing is probably still a
good idea.

BUG: 29365
Test: run wlan-smoke-test, run wlan-stress-test on astro, vim2
Change-Id: I9312166b90fce88b085caa9c7ac602e19d44e8c5