[fs-management] convert mount and fmount to use new functions
instead of launching the filesystem process themselves, the mount and
fmount functions now use the fs_init, fs_register, and fs_root_handle
functions to launch the filesystem using the export directory model. the
behavior change is minimal - the primary difference is that any
filesystems launched with mount are automatically registered with the
fshost registry.

Test: /system/test/fs/fs-management-test
Bug: 34530
Change-Id: Ifee36be9c63560b828ac9c4662d5aaf9aadb618d