[blobfs] move export directory creation to runner
While implementing the admin service, I realized that the primary use
case of the service (shutting down the filesystem) is done by the runner
and not by the blobfs class itself. In fact, I think it makes sense for
the runner to in general be responsible for running the set of related
filesystem services, and not the blobfs class itself.  This is a
relatively straightforward refactor to that effect.

Change-Id: Ib316ffeb0a08ca2271585b649bc0ee7e31e486e5