[debugger] Make "bt" not print registers by default.
There are two commands to print backtraces, "f" which prints just the
function name and line number, and "f -v"/"bt" which lists function
parameters and stack pointer information.

The stack pointer information is rarely needed and makes the "bt" output
more complex. The stack pointers are moved behind a "-v" flag on the
backtrace command (still the same as "f -v"), while "bt" by itself lists
only the function parameters.

Minor cleanup to formatting options. The variable formatting options for
showing variables are now piped through FormatFrameList() which gives
more control to the caller. Previously these were hardcoded inside
FormatFrameList().

Whether to force sync the frame list is now passed in as an option to
FormatFrameList(). Previously it always updated the frame list which the
callers weren't expecting and they would also invalidate the frame list
(always syncing was added later). Adding this flag (which is currently
always set to true) prevents the situation where the frames are empty
during the period of the call (in case there is something executing
asynchronously during that time). The command options to force cyns the
list are removed since they didn't do anything.

Change-Id: I3a5eb903f42bad76610b996a728ff95c9fab5f27