[lockdep] Avoid recursive ThreadLock
If a loop detection pass is triggered whilst the thread_lock is held
then the call to event_signal will attempt to acquire the already
acquired thread_lock, which will panic.

This change attempts to detect if the thread_lock is held, and use the
event_signal_thread_locked if it is. Additional plumbing for AssertHeld
on the thread_lock is added to achieve this.

Change-Id: Icc30b6975f628efc70c3e3ebb77fb18cc2d3481d