[componentmgr] Reduce identifier scopes
Changes to how model/mod.rs exposes identifiers:

- Currently, model/mod.rs publicly re-exports identifiers from
sub-modules in its own namespace. Instead, require clients to explicitly
qualify module holding the identifier. This is cleaner and more
explicitly describes which modules we're making public.

- Many identifiers are overly exposed. Some `pub` identifiers in
sub-modules only need to be `pub(super)` or `pub(crate)`, so that they
are usable from within model/crate but not outside of it.

- Remove globs, in favor of explicit imports, which are easier to
maintain and refactor. (Some globs remain, from external and testing
libraries.)

- Move `breakpoints` out of `testing` and `model`. Breakpoints are now
part of component manager in debug mode, so they are not test-only.

Change-Id: I18c2ae64a31a8fbd1337100193f787b112d539e0