[cobalt] Consume HTTP response headers from Clearcut
Previously we have been ignoring the HTTP response headers returned from
the request to Clearcut, even though Cobalt Core's code has a place to
put them. (Cobalt core doesn't currently do anything with these response
headers other than write them into a log error message in the case
that the response code is other than 200 OK.)

With this CL we now consume the headers and copy them to the appropriate
place in the Cobalt Core data structures.

We also make some improvements to fuchsia_http_client_test.

- We have eliminated the subclass of the class under test and
the associated use of condition variables becasue these are
not necessary to test the correctness of the code under test.

- We add tests of the new response header functionality.

- We added tests with longer response bodies.

- We simplified the logic in PrepareResponse() and removed the
need to wait on a condition variable. It is possible this was
the cause of earlier test flakes.

- We added a main() function to allow us to see VLOG messages
from the code under test.

Change-Id: I6ff0eca6a5e7508128edf54d2816ed53668888fd