[camera][controller] Keep track of the buffers handed off to child nodes
Consider a scenario in which a parent node (such as GDC) has two child
nodes. The parent must wait for both of its children to release a buffer
before we ask the parent's driver (GDC in this example) to return the
buffer to the free pool.

This CL adds the necessary book keeping to support this scenario.

Tests: This functionality is tested in CLs up in the stack, due to
    feature of having multiple streams is missing at this point.
    1. TestInUseBufferCountValidation()
    2. Integration test to validate all streams of monitor stream
    config.
    Ran fx test camera_tests to ensure no existing test cases break.

Change-Id: Ib6be431b476988a17ac5962c442ac1a647bae985