[i18n] Fix the handling of tzdata in icu_data::Loader
It turns out that whenever tzdata is unloaded from the ICU library, it is
necessary to call `u_cleanup` to remove any stale references to the removed
data before another data load can be attempted.  Not doing this meant
there were internal ICU caches that still had references to unloaded data
when a Loader went out of scope.

The fix is to add a call to `u_cleanup` in UDataMemory::drop().

The tests in icu_data tested telescoping loaders, but did not test what
happens if a loader goes out of scope *and* an unload happens before
another load is requested.  Adding that as a regression test teased
the issue out.

Bug: 42816
Change-Id: I487315ddc50e89970f6f0c09e685beae35443431