[pkg] Add optional root_threshold to RepositoryConfig
This allows the threshold of keys to be specified, rather than
being hardcoded to 1. We might want this if we decide to sign the root
metadata with a multiple offline keys.

Note: that this does not yet change the resolver to specify the config
threshold because that's blocked behind a TUF roll.

Change-Id: Idadd972b271d20654dd10c84e05eb284ef729229