[graphics][compute] Add a mock Spinel implementation for testing.
This CL introduces mock_spinel::Spinel, a mock implementation of
the Spinel API (sans spn_context_create() and spn_render()) that
can be used to properly test any kind of code that calls the
Spinel API. For example, this will be used in a future CL to
change the spinel2svg_unittest, removing dependencies to internal
Spinel details.

Another CL will use it to implement a Skia-based implementation
of the Spinel API for demos and comparison purpose.

NOTE: This requires moving the svg2spinel unittest into a separate
      test package, since it currently depends on Spinel internals
      and requires linking against it. A future CL will refactor
      the test to avoid this.

Bug: None
Test: fx run-test graphics_compute_unittests --gtest_filter=MockSpinelTest.*
      fx run-test graphics_compute_spinel_unittests

Change-Id: I4a50c81740bb2a9166abb310b4b0a56bc99b37c0