[wlan][policy] Add in memory network configs
Save and use network configs in memory with saved network manager where
we used known ess before. Continue to store to KnownEssStore and load
network configs from this location in order to preserve persistence. SSID
and password are the only meaningful fields in the configs so far.

Bug: 35908
Tests: Modified existing tests:
       fx run-test wlancfg-tests
       confirmed behavior is unchanged through manual testing, setting
       up connections.
Change-Id: I618a2594c204a39938d24124be6249b7cf7316c5