[fidl][fuzzing] Turn off debugging by default.
Currently the generated code for FIDL fuzzing includes a number of
diagnostics messages. This can be useful when developing a new FIDL
fuzzer, but should generally be omitted.

This CL adds a simple macro to disable printing if a flag is not set.
The flag can be set and printing re-enabled simply by including
`configs += [ "//build/fuzzing:verbose_logging" ] in the GN target.

Bug: 42007
Change-Id: Id898f6760b7d4a7f32d2db46ee6cdcc8fb513b8e