[kernel][sched] Fix edge case in nested duration tracing.
Fixes an edge case in nested duration tracing in the scheduler.
Durations need to be completed before performing a context switch,
otherwise the trace events in the buffer can become unmatched.

Previously, this was addressed by passing a void pointer to the outer
duration into RescheduleCommon. RescheduleCommon automatically ends
the passed in duration if a reschedule occurs.

The edge case arises if some of the TraceDuration instances have
different types, due to changing their template parameters. The void
pointer loses this information.

For example:

  TraceDuration<..., TraceContext::Cpu>

And

  TraceDuration<..., TraceContext::Thread>

These have different types and result in different bits written to
the trace buffer. This change addresses the problem using fbl::Function
to capture both the trace instance and the appropriate action to take
to end the duration correctly.

Test: Duration traces with varying parameters are properly terminated.
Change-Id: I110615facad203b287077607bf8e281d373a12df