[camera][ge2d] CropRectangle name cleanup
This is a minor change to simplify the naming of the crop rectangle in
the GE2D interface. The rectangle in question is nothing more than a
standard Rect w/unsigned values for x, y, width, and height.

This CL also cleans up some whitespace & line length violations.

Tested: Verified compilation. The change is only a name change and does
not alter functionality.

Change-Id: Ib3a7d3c08de640fee918b681f5ef6c20cf5d8fcc