[fidl] Fix designated initializer use in transformer.cc
This was accidentally using a combination of designated initializer and
position based initialization. This is valid-if-confusing C, but isn't
going to be valid C++ (when designated initializers are in a later C++
standard) and so some warning settings will complain, e.g.:

https://logs.chromium.org/logs/crashpad/buildbucket/cr-buildbucket.appspot.com/8894243386077508688/+/steps/compile_with_ninja/0/stdout

We could also disable that warning in the Crashpad build, since I guess
we do in-tree, but since it seems wrong anyway, we should probably just
fix it.

Change-Id: I095cb30b9ce12aee9dbb4629605e43a86f44d0ee