[drivers] Prefix iommu.enable as driver option
Changes the iommu.enable option to be under the driver 'namespace' of
options.

Motivation for this change is that only options prefixed with 'driver.'
are passed on to devhosts and without this change it actually is not
possible to enable the hardware iommu in the x86 platform driver.

Test: Added 'driver.iommu.enable=true' to kernel cmdline and validated
use_hardware_iommu returned true

Change-Id: I2cc34b3a8c40129688928aae86111b2bd5e564d5