[camera] Switch to hw_accel_frame_callback (step 2)
This CL is one of several that transitions the ISP interface from
using OutputStreamCallback to using the hw_accel_frame_callback
that is common to other camera-related hardware accelerators.
The hw_accel_frame_callback provides additional information about
a frame, including its metadata and status. The changes also move the
name of CreateOutputStream2 back to its original CreateOutputStream.

This CL adds CreateOutputStream() to isp.banjo.

Testing: Re-ran all the camera tests, as this CL only adds a function
that has yet to be called.

Change-Id: Ida18b8881c5b3e8593dec72f333cec02764758be