[cm] Update ComponentRunner protocol to return an error.
Currently, the ComponentRunner `Start` method is "fire and forget", with
no mechanism to return an error. The component manager, however, may
wish to know about failures of a component to launch, without
neccessarily having the channel close.

Update the protocol to make `Start` return an error code, and update the
component manager code to handle this error (which, at the moment, is
simply logging it.)

Bug: 4761

Change-Id: I241c28ddf9f57888a1e7d8847f5a39dbb70cce97