[cm] Avoid hard-coding the builtin set of runners.
Currently, the CM startup code assumes there exists precisely one
builtin runner named "elf" which is given to the root component.
This complicates testing, which may not want to make the same
assumptions.

This CL adjusts model / root realm creation to accept an arbitrary
set of builtin runners. The singleton "elf_runner" still remains
in the short term, which continues to be used as the default runner
until we can remove the assumption that a default runner exists.

Bug: 4761
Change-Id: I437756b88e010d8ed821ab8d5c959d4da596b89b