[cm] Update binding tests to explicitly use and route ELF runner caps.
Upcoming changes will require that the ELF runner capability be
explicitly used and offered to children. This change updates the tests
in "binding.rs" to avoid assuming that ELF programs are run by default.

Modifying the in-test ComponentDecl's explicitly causes them to become
quite verbose. To reduce the verbosity, this CL also adds
a ComponentBuilder struct, which reduces some of the boilerplate of
creating a ComponentDecl.

The builder can be instructed to route the ELF capabilities from the
containing realm to all children with the call
"offer_elf_runner_to_children". The hope is that this is a good balance
between being too verbose (e.g., explicitly having to offer the cap to
every child) and being too magic (e.g., silently routing the cap in the
background, with no indication that this is happening inside the test.)

Bug: 4761
Change-Id: Id2cd3473ad3d72cd4dcbfcfae0df5e8d308ed014