[camera][calibrate] SL4F Facade
This CL provides the SL4F Facade for communication between the factory
host and the on-device FIDL service. The Facade will be incorporated into
SL4F proper in a future CL. The corresponding FIDL interface is
fxr/332809.

Note: This CL only guarantees the correctness of de/serialization of
API responses. Many changes are expected to both this facade and the
FIDL interface. End-to-end functionality is implemented in followup CLs.

Tested: tests written in `types.rs` and `facade.rs`. Requires the
`--with garnet/bin/sl4f:service-tests` build flag and can be run with
`fx run-test sl4f-service-tests`

Change-Id: I867dd2eb7fd685b57502b7fc2f4efcb0acf946bc