[crash] database handles upload attempt counts
* The number of upload attempts for a crash reports will be logged in
Cobalt both while the report is pending and at the end of its life.
Because the database handles recording end of life metrics for reports
it's easier to have the database manage counting all upload attempts
instead of the queue counting upload attempt while the report is pending
and the database counting the number of upload attempts when a report is
no longer pending.

Bug: 42775

TESTED=`fx run-test crashpad_agent_tests`

Change-Id: I26b30932ec2b9cc6f4c988b9402b261c450a0d78