[benchmarks] Add SL4F wrapper for rust_inspect_benchmarks
Convert rust_inspect_benchmarks to use package() instead of
test_package().

That involves:
 * change "tests" field to "binaries"
 * reference "bin/" instead of "test/" in the .cmx
 * add "meta" field to add the .cmx (that was being added implicitly
   when this was a test_package)
 * update README.md for how to run the test

That is a workaround for the following problem: Adding the
test_package() as a GN dependency for SL4F tests causes it to be run
as a unit test under QEMU on the terminal-x64-release builder, but
running Fuchsia on QEMU fails on that builder.  (See comments in
garnet/bin/zircon_benchmarks/BUILD.gn where I used a different
workaround for this problem.)

This means we no longer run rust_inspect_benchmarks in unit test mode
on the bots, but that's not a big deal.

Bug: 23091
Test: on CQ: terminal-x64-release trybot;
  local testing: "fx set terminal.x64", "fx run-e2e-tests rust_inspect_benchmarks_test"
Change-Id: I3ae3915a689c49a2e85bff7e5de1b955568fa42e