[gdc][camera] Take locks for all GDC external APIs
The GDC driver can serve multiple clients. This means that the driver
needs to grab locks for all the APIs exposed by the GDC banjo interface in
order to avoid race conditions.

This change also adds another entry to the TaskInfo structure, which is the
task index. This will be used in future CLs to remove tasks from the
processing queue.

Test: No functionality changes here, just adding locks.
      Ran fx test camera_tests successfully.

Change-Id: I276d7a1a599d4e573671d0ff15b109d6ec53a45d