[rust] Add rust_panic build arg
This adds support for setting rust_panic="abort" in our build. It leaves
the default as "unwind" as it is today.

When using rust_panic="abort", we need a way to add the
sys.process.Launcher service to every test's sandbox. Rust's test
framework uses this to launch a subprocess for every unit test.

We do this by collecting additions to our test package manifest using
metadata. This required a change to `cmc merge` to allow an array at the
root of a file to be merged. In this case, all elements of the array
will get merged into the final object.

Not everything builds/tests successfully with this arg, but those will
be fixed in follow up changes.

Part of fxb/26968

Change-Id: Idba86d15b44782d03624c006bb5d61f20d4a130a