[debugger] Allow InputLocations to be saved in settings.
This is not currently used and there is no behavior change.

In the future it will be used to convert BreakpointSettings to a
SettingStore and use get/set for breakpoint options.

TEST: Included unit tests. There is no current test for the
verbs_settings change because there is no current setting with this type
(these tests use the real schemas). That will be added when we convert
breakpoints to use this new schema.

Change-Id: Ifad8f97fd0d0f19fda7b7eba7b9fb8e8b2ea27a1