[camera][controller] Implement OnShutdown() for ProcessingNode class
This CL adds support for the OnShutdown() method which is called
upon clients disconnecting a stream.
Depending on what type of node it is, the controller would
request ISP, GDC and GE2D to free up resources pertaining to this task.

This CL fixes the memory leaks happening in GDC driver only.

NOTE: This CL shuts down the stream partially. Specifically the GDC
driver only frees up the resources but does not handle existing
queued up tasks in the queue and the tasks which are currently
processing. This will sometimes lead to a crash upon disconnecting the
camera demo. A complete solution is still work in progress.

Fixed: 42441

Test: fx test camera_tests
      Ran demo on Sherlock device multiple times to ensure no memory
      leaks happen.
      Ran buffer leaks tests to ensure no leaks are happening.

Change-Id: I6ad4bc8657f56d7bcffecb61d7e61e0ea1f64cfa