[a11y][3/5] Implements Semantic Tree Service.
This cl is part of an ongoing refactor of accessibility Semantics manager.

This cl implements the fuchsia service responsible for dealing with semantics
updates.

A bunch of tests could be deleted as they are now being implemented on the base
SemanticTree class.

The services (Semantics manager service and Semantic Tree Service), now only
have to deal with server-side stuff, leaving the logic of tree manipulation
details to the Semantic Tree data class.

TEST:
    fx run-test a11y_lib_tests

Change-Id: I3416fe32e87415968b1bd25de902fb0d6ce9aa8f