[timezones] ICU data extractor host tool
Replaces extract_revision.sh with a host tool implemented in C++,
icu_data_extractor, which uses the ICU4C library.

The tool can easily be extended in the near future, e.g. to
extract a list of time zone IDs for workstation settings at build
time.

Testing: This is a build tool; if it doesn't behave as expected, the
build will fail.

Bug: 42549
Bug: 42827
Bug: 42861

Change-Id: I5d41b34b311fe6bf33db4a30c9d3a2083eeb5068