[bt][hci-uart] Fix possible use-after-free.
serial_write now takes ownership of the buffer it's passed (for asynch
serial).  Previously the snoop log was written to after serial_write,
which may not be valid any longer, as the asynchronous serial callback
frees the same buffer when it is finished writing.

log to the snoop channel before actually writing to serial to eliminate
this possibility.

Add locking to the write in hci_handle_acl_reasd_events as it was
missing before and this function is called while unlocked.

Bug: 42413
Change-Id: Ifba2da8f669dea6f0c4b3ef4664f43106fd9dde3