[benchmarking][sl4f] Fix flutter frame stats metric logic
In the flutter frame stats metrics processor, handle the case in which a
frame group is connected to 0 "VSYNC" events.

Previously, we would throw an exception when trying to dereference
"followingVsyncs.first".  Now, emit a warning and move on to the next
frame group.

This will address the build stability issues found in 42964, however we
should also look into the root cause of this issue, which in 42964 was
the Scenic process running out of trace buffer space early, which caused
a break in the "vsync callback" -> "VSYNC" flow sequence.

Bug: 42964

Change-Id: I637ce183162d27e6a808b56d7d9772ce1dc2c33a